% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFR.R
\docType{package}
\name{BFR-package}
\alias{BFR}
\alias{BFR-package}
\title{BFR}
\description{
Bayesian Functional Regression
}
\details{
More details about the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frahik/BGFRA}
  \item Report bugs at \url{https://github.com/frahik/BGFRA/issues/new}
}

}
\author{
\strong{Maintainer}: Francisco Javier Luna-Vazquez \email{frahik@gmail.com}

Authors:
\itemize{
  \item Osval Antonio Montesinos-Lopez \email{oamontes1@ucol.mx}
}

}
