% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGFRA.R
\name{BGFRA}
\alias{BGFRA}
\title{BGFRA}
\usage{
BGFRA(y, response_type = "gaussian", a = NULL, b = NULL, ETA = NULL,
  nIter = 1500, burnIn = 500, thin = 5, saveAt = "", S0 = NULL,
  df0 = 5, R2 = 0.5, weights = NULL, verbose = FALSE,
  rmExistingFiles = TRUE, groups = NULL, Folds = 0, set_seed = NULL)
}
\arguments{
\item{y}{\code{data.frame} the data with $Response, $Line and $Env, defined on it (NAs allowed).}

\item{response_type}{\code{string} can be "gaussian" or "ordinal"}

\item{a, b}{\code{numeric} .}

\item{ETA}{\code{list} two level list.}

\item{nIter, burnIN, thin}{\code{integer}}

\item{saveAt}{\code{string}}

\item{S0, df0}{\code{numeric}}

\item{R2}{\code{numeric}}

\item{weights}{\code{numeric}}

\item{verbose}{\code{logical}}

\item{rmExistingFiles}{\code{logical}}

\item{groups}{\code{factor}}

\item{Folds}{\code{integer}}

\item{set.seed}{\code{integer} to reproducible research}
}
\description{
Bayesian Genomic Functional Regression Analysis
}
\details{
This is a generic function
}
