% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{boxplot.BFRCV}
\alias{boxplot.BFRCV}
\title{boxplot.BFRCV}
\usage{
\method{boxplot}{BFRCV}(x, select = "Pearson", ordered = T, ...)
}
\arguments{
\item{x}{\code{BFRCV object} Objeto BFRCV, resultado de ejecutar BFR() con el parametro folds > 2}
}
\description{
Solo es una prueba
}
