% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{boxplot.GFRCV}
\alias{boxplot.GFRCV}
\title{boxplot.GFRCV}
\usage{
\method{boxplot}{GFRCV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{\code{GFRCV object} Objeto GFRCV, resultado de ejecutar GFR() con el parametro folds > 2}
}
\description{
Solo es una prueba
}
