% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{boxplot.MTMECV}
\alias{boxplot.MTMECV}
\title{boxplot.MTMECV}
\usage{
\method{boxplot}{MTMECV}(x, select = "Pearson", ordered = TRUE, ...)
}
\arguments{
\item{x}{\code{MTMECV object} Objeto MTMECV, resultado de ejecutar MTME()}

\item{select}{\code{string} Pearson or MSEP}

\item{ordered}{\code{logic} TRUE or FALSE}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
