% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_Bspline.R
\name{Bspline.Basis}
\alias{Bspline.Basis}
\title{Bspline.Basis}
\usage{
Bspline.Basis(Bands, Wavelengths = NULL, nBasis = 1, interaction = NULL)
}
\arguments{
\item{Bands}{(\code{matrix}) Lorem ip sum}

\item{Wavelengths}{(\code{integer}) Lorem ip sum}

\item{nBasis}{(\code{integer}) Lorem ip sum}

\item{interaction}{(\code{matrix}) Lorem ip sum}
}
\description{
Lorem ip sum
}
