% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{cleanDat}
\alias{cleanDat}
\title{cleanDat function}
\usage{
cleanDat(forceClean = F)
}
\arguments{
\item{forceClean}{By default is FALSE, else, the function not requiere the confirmation by the user to delete de .dat files.}
}
\value{
logical vector, of the files deleted.
}
\description{
Easy way to delete dat files
}
