% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{crossvalidation}
\alias{crossvalidation}
\title{Cross-Validation with K Folds}
\usage{
crossvalidation(data, K = 5, set_seed = NULL)
}
\arguments{
\item{data}{\code{data.frame} Object with the $Response, $Line and $Env especified on it.}

\item{K}{\code{integer} Number of groups to the cross-validation.}

\item{set_seed}{\code{integer} Number of seed for replicable research.}
}
\description{
Example
}
