% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.KFold}
\alias{CV.KFold}
\title{Cross-validation with K Folds}
\usage{
CV.KFold(data, DataSetID = 'Line', K = 5, set_seed =  123)
}
\description{
Cross-validation with K Folds
}
\keyword{internal}
