% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.RandomPart}
\alias{CV.RandomPart}
\title{Cross-validation with Random partitions}
\usage{
CV.RandomPart(data, Npartitions = 10,  PTesting = 0.2, set_seed = 123)
}
\description{
Cross-validation with Random partitions
}
\keyword{internal}
