% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.Stratified}
\alias{CV.Stratified}
\title{Cross-Validation with stratified samples}
\usage{
CV.Stratified(DataSet, NSamples = 10, nTesting = 10, replace = FALSE, set_seed = NULL)
}
\description{
Cross-Validation with stratified samples
}
\keyword{internal}
