% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CV.StratifiedByFrac}
\alias{CV.StratifiedByFrac}
\title{Cross-validation with stratified samples}
\usage{
CV.StratifiedByFrac(DataSet, NSamples = 10, fracTesting = 0.1, replace = FALSE, set_seed = NULL)
}
\description{
Cross-validation with stratified samples
}
\keyword{internal}
