% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{doubleinteractionMatrix}
\alias{doubleinteractionMatrix}
\title{double Interaction Matrix}
\usage{
doubleinteractionMatrix(X1, X2)
}
\arguments{
\item{X1}{\code{matrix}}

\item{X2}{\code{matrix}}
}
\description{
example
}
