% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{Fourier.Basis}
\alias{Fourier.Basis}
\title{Fourier.Basis}
\usage{
Fourier.Basis(Bands, Wavelengths = NULL, n.basis = 1, interaction = NULL)
}
\arguments{
\item{Bands}{\code{matrix}}

\item{n.basis}{\code{integer}}

\item{interaction}{\code{matrix}}

\item{Wavelenghts}{\code{vector}}
}
\description{

}
