% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplicitImports.R
\name{getMatrixForm}
\alias{getMatrixForm}
\title{Get the matrix format from a tidy format}
\usage{
getMatrixForm(data, onlyTrait = T)
}
\description{
Get the matrix format from a tidy format
}
\keyword{internal}
