% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplicitImports.R
\name{getTidyForm}
\alias{getTidyForm}
\title{Get the tidy format from a matrix format}
\usage{
getTidyForm(data, onlyTrait = T)
}
\description{
Get the tidy format from a matrix format
}
\keyword{internal}
