% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFR.R
\docType{package}
\name{GFR-package}
\alias{GFR}
\alias{GFR-package}
\title{GFR}
\description{
Bayesian Functional Regression
}
\details{
More details about the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frahik/GFR}
  \item Report bugs at \url{https://github.com/frahik/GFR/issues/new}
}

}
\author{
\strong{Maintainer}: Francisco Javier Luna-Vazquez \email{frahik@gmail.com}

Authors:
\itemize{
  \item Osval Antonio Montesinos-Lopez \email{oamontes1@ucol.mx}
}

}
