% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maize_BFR.R
\docType{data}
\name{Maize_BFR}
\alias{Maize_BFR}
\title{Maize Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\usage{
data('Maize_BFR')
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
}
