% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTME.R
\name{MTME}
\alias{MTME}
\title{Multi-trait Multi-Environment}
\usage{
MTME(data = NULL, ETA = NULL, nIter = 2500, burnIn = 500, thin = 5,
  progressBar = TRUE, CrossValidation = NULL, set_seed = NULL,
  digits = 4, ...)
}
\arguments{
\item{data}{data}

\item{ETA}{eta}

\item{nIter}{number of iterations}

\item{burnIn}{number of burning}

\item{thin}{number of thinning}

\item{CrossValidation}{cv object}

\item{set_seed}{seed for a reproducible research}

\item{digits}{number of digits of accuracy in the results}

\item{...}{for BFR function}
}
\value{
Something cool
}
\description{
Multi-trait Multi-Environment
}
\examples{
ETA=list(Env=list(X=Z.E,model="BRR"),Gen = list(X = Z.G, model = 'BRR'), EnvGen=list(X=Z.EG,model="BRR"))
CrossValidation <- BMTME::CV.RandomPart(pheno, NPartitions = 10, PTesting = 0.2, set_seed = 123)
}
