% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot.BFR}
\alias{plot.BFR}
\title{plot.BFR}
\usage{
\method{plot}{BFR}(x, ...)
}
\arguments{
\item{x}{\code{BFR object} Objeto BFR, resultado de ejecutar BFR}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
