% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{plot.BFRCV}
\alias{plot.BFRCV}
\title{Plot BFRCV graph}
\usage{
\method{plot}{BFRCV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{BFRCV object} BFRCV object, result of use the BFR() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the BFR Object, else ('MSEP'), plot the MSEP of the BFRCV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BFRCV object
}
