% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.BGFRA}
\alias{plot.BGFRA}
\title{plot.BGFRA}
\usage{
\method{plot}{BGFRA}(x, ...)
}
\arguments{
\item{object}{\code{BGFRA object} Objeto BGFRA, resultado de ejecutar BGFRA}
}
\description{
Solo es una prueba
}
