% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot.BGFRACV}
\alias{plot.BGFRACV}
\title{Plot BGFRACV graph}
\usage{
\method{plot}{BGFRACV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{BGFRACV object} BGFRACV object, result of use the BGFRA() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the BGFRACV Object, else ('MSEP'), plot the MSEP of the BGFRACV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from BGFRACV object
}
