% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot.GFRCV}
\alias{plot.GFRCV}
\title{Plot GFRCV graph}
\usage{
\method{plot}{GFRCV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{GFRCV object} GFRCV object, result of use the GFR() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the GFR Object, else ('MSEP'), plot the MSEP of the GFRCV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from GFRCV object
}
