% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{plot.MTMECV}
\alias{plot.MTMECV}
\title{Plot MTMECV graph}
\usage{
\method{plot}{MTMECV}(x, select = "Pearson", ...)
}
\arguments{
\item{x}{\code{MTMECV object} MTMECV object, result of use the MTME() function}

\item{select}{\code{character} By default ('Pearson'), plot the Pearson Correlations of the MTME Object, else ('MSEP'), plot the MSEP of the MTMECV Object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Plot from MTMECV object
}
