% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{predict.BFR}
\alias{predict.BFR}
\title{predict.BFR}
\usage{
\method{predict}{BFR}(object, newdata, ...)
}
\arguments{
\item{object}{\code{BFR object} Objeto BFR, resultado de ejecutar BFR}
}
\description{
Solo es una prueba
}
