% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.BGFRA}
\alias{predict.BGFRA}
\title{predict.BGFRA}
\usage{
\method{predict}{BGFRA}(object, newdata, ...)
}
\arguments{
\item{object}{\code{BGFRA object} Objeto BGFRA, resultado de ejecutar BGFRA}
}
\description{
Solo es una prueba
}
