% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{predict.GFR}
\alias{predict.GFR}
\title{predict.GFR}
\usage{
\method{predict}{GFR}(object, newdata, ...)
}
\arguments{
\item{object}{\code{GFR object} Objeto GFR, resultado de ejecutar GFR}
}
\description{
Solo es una prueba
}
