% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{print.BFRCV}
\alias{print.BFRCV}
\title{Print BFRCV information object}
\usage{
\method{print}{BFRCV}(x, ...)
}
\arguments{
\item{x}{object a}

\item{...}{more objects}
}
\value{
test
}
\description{
Print BFRCV information object
}
