% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.BGFRA}
\alias{residuals.BGFRA}
\title{residuals.BGFRA}
\usage{
\method{residuals}{BGFRA}(object, ...)
}
\arguments{
\item{object}{\code{BGFRA object} Objeto BGFRA, resultado de ejecutar BGFRA}
}
\description{
Solo es una prueba
}
