% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{residuals.GFR}
\alias{residuals.GFR}
\title{residuals.GFR}
\usage{
\method{residuals}{GFR}(object, ...)
}
\arguments{
\item{object}{\code{GFR object} Objeto GFR, resultado de ejecutar GFR}
}
\description{
Solo es una prueba
}
