% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{simpleinteractionMatrix}
\alias{simpleinteractionMatrix}
\title{Simple Interaction Matrix}
\usage{
simpleinteractionMatrix(X)
}
\arguments{
\item{X}{\code{matrix}}
}
\description{
example
}
