% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.BFR}
\alias{summary.BFR}
\title{Summary.BFR}
\usage{
\method{summary}{BFR}(object, ...)
}
\arguments{
\item{object}{\code{BFR object} Objeto BFR, resultado de ejecutar BFR}
}
\description{
Solo es una prueba
}
