% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.BFRCV}
\alias{summary.BFRCV}
\title{Summary.BFRCV}
\usage{
\method{summary}{BFRCV}(object, ...)
}
\arguments{
\item{object}{\code{BFRCV object} Objeto BFRCV, resultado de ejecutar BFR}
}
\description{
Solo es una prueba
}
