% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.BGFRA}
\alias{summary.BGFRA}
\title{Summary.BGFRA}
\usage{
\method{summary}{BGFRA}(object, ...)
}
\arguments{
\item{object}{\code{BGFRA object} Objeto BGFRA, resultado de ejecutar BGFRA}
}
\description{
Solo es una prueba
}
