% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.GFR}
\alias{summary.GFR}
\title{Summary.GFR}
\usage{
\method{summary}{GFR}(object, ...)
}
\arguments{
\item{object}{\code{GFR object} Objeto GFR, resultado de ejecutar GFR}
}
\description{
Solo es una prueba
}
