% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.GFRCV}
\alias{summary.GFRCV}
\title{Summary}
\usage{
\method{summary}{GFRCV}(object, ...)
}
\arguments{
\item{object}{\code{GFRCV object} GFRCV object, result of use the GFR() function}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of GFRCV object
}
