% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{summary.MTMECV}
\alias{summary.MTMECV}
\title{Summary.MTMECV}
\usage{
\method{summary}{MTMECV}(object, information = "compact", digits = 4, ...)
}
\arguments{
\item{object}{\code{MTMECV object} Objeto MTMECV, resultado de ejecutar MTME()}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
