% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{tripleInteractionMatrix}
\alias{tripleInteractionMatrix}
\title{triple Interaction Matrix}
\usage{
tripleInteractionMatrix(X1, X2, X3)
}
\arguments{
\item{X1}{\code{matrix}}

\item{X2}{\code{matrix}}

\item{X3}{\code{matrix}}
}
\description{
example
}
