% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_BFR.R
\docType{data}
\name{Wheat_BFR}
\alias{Wheat_BFR}
\title{Wheat Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\usage{
data('Wheat_BFR')
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
}
