% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheat_BGFRA.R
\docType{data}
\name{wheat_BGFRA}
\alias{wheat_BGFRA}
\title{Wheat Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\usage{
data(wheat_BGFRA)
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
}
