% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_GFR.R
\docType{data}
\name{Wheat_GFR}
\alias{Wheat_GFR}
\title{Wheat Data}
\format{a \code{TidyData} format, 3000 row per 4 columns.}
\usage{
data('Wheat_GFR')
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
http://sandsynligvis.dk/articles/18/codebook.html
}
