% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WheatI_GFR.R
\docType{data}
\name{WheatI_GFR}
\alias{WheatI_GFR}
\title{Wheat Data}
\usage{
data('WheatI_GFR')
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
}
