% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boxplot.BGFRACV}
\alias{boxplot.BGFRACV}
\title{boxplot.BGFRA-CV}
\usage{
\method{boxplot}{BGFRACV}(x, ordered = F, ...)
}
\arguments{
\item{object}{\code{BGFRA-CV object} Objeto BGFRA-CV, resultado de ejecutar BGFRA() con el parametro folds > 2}
}
\description{
Solo es una prueba
}
