% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bspline.R
\name{Bspline.Basis}
\alias{Bspline.Basis}
\title{Bspline.Basis}
\usage{
Bspline.Basis(Bands, Wavelengths = NULL, n.basis = 1, interaction = NULL)
}
\arguments{
\item{Bands}{\code{matrix}}

\item{n.basis}{\code{integer}}

\item{interaction}{\code{matrix}}

\item{Wavelenghts}{\code{vector}}
}
\description{
Solo es una prueba
}
