% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{crossvalidation}
\alias{crossvalidation}
\title{Cross-Validation}
\usage{
crossvalidation(dat_F, K = 5, set_seed = NULL)
}
\arguments{
\item{dat_F}{\code{data.frame}}

\item{K}{\code{integer}

#rm(list=ls(all=TRUE))
#K Numero de grupos
#dat_F$Line and dat_F$Env}
}
\description{
Example
}
