% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA.R
\name{ETAGenerate}
\alias{ETAGenerate}
\title{ETAGenerate}
\usage{
ETAGenerate(dataset, datasetID = "Line", Bands = NULL, Wavelengths = NULL,
  n.basis = 1, functionalType = Fourier.Basis, priorType = "FIXED",
  bandsType = "Alternative")
}
\arguments{
\item{dataset}{TidyFormat}

\item{datasetID}{column with the identifiers.}

\item{Bands}{Bands}

\item{Wavelengths}{Wavelenths}

\item{n.basis}{Number of basis by default only use 1 basis.}

\item{functionalType}{Basis function, by default is Fourier.Basis also could be Bspline.Basis.}

\item{priorType}{Prior to assign, by default is 'FIXED', could be 'BRR', 'BayesA', 'BayesB', 'BayesC' or 'BL'}

\item{bandsType}{Model to apply in bands, by default 'Alternative' will be used, also could be 'Simple', 'Complex'}
}
\description{
Function to generate a Linear Predictor from a dataset to BFR.
}
