% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA.R
\name{ETAGenerate}
\alias{ETAGenerate}
\title{ETAGenerate}
\usage{
ETAGenerate(dataset, datasetID = "Line", Multivariate = "Traditional",
  GenomicMatrix = NULL, REML = NULL, priorType = "BRR", Bands = NULL,
  Wavelengths = NULL, method = "Simple", basisType = "Fourier.Basis",
  nBasis = 1, ...)
}
\arguments{
\item{dataset}{TidyFormat}

\item{datasetID}{column with the identifiers.}

\item{Multivariate}{By default, when de dataset includes more than one Environment and more than one Trait (MTME) the solution is adjusted by the method traditional, also is possible adjust the MTME by "SVD" <doi: >}

\item{GenomicMatrix}{lalalalal}

\item{REML}{Logical, By default is NULL, If is TRUE, the priorType is ommited else If is FALSE, the priorType is ommited and.}

\item{priorType}{Prior to assign, by default is 'FIXED', could be 'BRR', 'BayesA', 'BayesB', 'BayesC' or 'BL'}

\item{Bands}{Bands}

\item{Wavelengths}{Wavelenths}

\item{method}{Model to apply in bands, by default 'Alternative' will be used, also could be 'Simple', 'Complex'}

\item{basisType}{Basis function, by default is Fourier.Basis also could be Bspline.Basis.}

\item{nBasis}{Number of basis by default only use 1 basis.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Function to generate a Linear Predictor from a dataset to GFR.
}
