% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis_Fourier.R
\name{Fourier.Basis}
\alias{Fourier.Basis}
\title{Fourier.Basis}
\usage{
Fourier.Basis(Bands, Wavelengths = NULL, nBasis = 1, interaction = NULL)
}
\arguments{
\item{Bands}{(\code{matrix}) Lorem ip sum}

\item{nBasis}{(\code{integer}) Lorem ip sum}

\item{interaction}{(\code{matrix}) Lorem ip sum}

\item{Wavelenghts}{(\code{integer}) Lorem ip sum}
}
\description{
Lorem ip sum
}
