% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFR.R
\docType{package}
\name{GFR-package}
\alias{GFR}
\alias{GFR-package}
\title{GFR}
\description{
Bayesian Functional Regression
}
\details{
BFR is an modificated version of BGLR which implements a Gibbs sampler for a Bayesian regression model, this new version allows to... More details about the package.
}
\seealso{
\code{\link[BGLR]{BGLR}}
}
\author{
\strong{Maintainer}: Francisco Javier Luna-Vazquez \email{frahik@gmail.com}

Authors:
\itemize{
  \item Osval Antonio Montesinos-Lopez \email{oamontes1@ucol.mx}
  \item Abelardo Montesinos-Lopez
  \item Jose Crossa
  \item Gustavo De los campos
}

}
