% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{predict.BFR}
\alias{predict.BFR}
\title{predict.BFR}
\usage{
\method{predict}{BFR}(object, newdata, ...)
}
\arguments{
\item{object}{\code{BFR object} Objeto BFR, resultado de ejecutar BFR}

\item{newdata}{\code{data.frame} Predictions}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
