% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{residuals.BFR}
\alias{residuals.BFR}
\title{residuals.BFR}
\usage{
\method{residuals}{BFR}(object, ...)
}
\arguments{
\item{object}{\code{BFR object} Objeto BFR, resultado de ejecutar BFR}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
