% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.BFRCV}
\alias{summary.BFRCV}
\title{Summary}
\usage{
\method{summary}{BFRCV}(object, ...)
}
\arguments{
\item{object}{\code{BFRCV object} BFRCV object, result of use the BFR() function}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Summary of BFRCV object
}
