% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{summary.BFRCV}
\alias{summary.BFRCV}
\title{Summary.BFRCV}
\usage{
\method{summary}{BFRCV}(object, information = "compact", digits = 4, ...)
}
\arguments{
\item{object}{\code{BFRCV object} Objeto BFRCV, resultado de ejecutar BFRCV()}

\item{information}{compact, extended, complete}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Solo es una prueba
}
