% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WheatI_GFR.R
\docType{data}
\name{WheatI_GFR}
\alias{WheatI_GFR}
\title{Wheat Data}
\format{\code{WheatI_GFR} is \code{data.frame} in \code{TidyData} format, 976 row per 3 columns. Also is loaded \code{WheatI_Wavelengths} and \code{WheatI_Bands} which includes the names of the bandas and the bands respectively.}
\usage{
data('WheatI_GFR')
}
\description{
The package includes a data set based on a portion of the data used in the study of
}
\author{
Montesinos-Lopez, O. A.
http://sandsynligvis.dk/articles/18/codebook.html
}
